quizzland-nextjs/backend/requirements.txt
appdirs==1.4.4
asgiref==3.3.4
boto3==1.17.99
botocore==1.20.99
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
coinex.py==1.0.4
colorama==0.4.4
coverage==5.5
distlib==0.3.1
dj-static==0.0.6
Django==3.2.11
django-admin-honeypot==1.1.0
django-ckeditor==6.1.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-js-asset==1.2.2
django-redis==5.0.0
django-robots==4.0
django-rq==2.4.1
django-storages==1.11.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.5.0
filelock==3.0.12
google-search-results==2.4.0
idna==2.10
jmespath==0.10.0
mysql-connector-python==8.0.25
mysqlclient==2.0.3
numpy==1.21.4
Pillow==9.0.0
playsound==1.2.2
protobuf==3.17.3
PyJWT==1.7.1
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
redis==3.5.3
requests==2.25.1
rq==1.8.1
s3transfer==0.4.2
selenium==3.141.0
six==1.16.0
sqlparse==0.4.2
static3==0.7.0
urllib3==1.26.5
virtualenv==20.1.0
